requirements.txt
Django>=3.2.0,<4.0.0
Pillow>=9.0.0
beautifulsoup4>=4.9.0
requests>=2.25.0
python-dotenv>=0.15.0
django-crequest>=2018.5.11
scour>=0.38.0
pillow-heif>=0.10.0
libsass>=0.21.0
django-debug-toolbar>=3.2.0
pytest>=6.2.0
pytest-django>=4.5.0
